+++ packages/terra-search-field/src/terra-dev-site/doc/example/SearchFieldDisableAutoSearch.jsx
@@ -1,11 +1,12 @@
 import React from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
-import SearchFieldExampleTemplate from 'terra-search-field/lib/terra-dev-site/doc/example/SearchFieldExampleTemplate';
+import SearchField from 'terra-search-field/lib/SearchField';
+
 
 const SearchFieldDisableAutoSearch = () => (
-  <SearchFieldExampleTemplate
-    disableAutoSearch
-  />
+  <div>
+    <SearchField disableAutoSearch />
+  </div>
 );
 
 export default SearchFieldDisableAutoSearch;